requirements.txt
# Deep Learning and ML
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Signal Processing
scipy>=1.11.0
wfdb>=4.1.0  # For MIT-BIH database
neurokit2>=0.2.0
PyWavelets>=1.4.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Real-time Processing
Flask>=2.3.0
Flask-SocketIO>=5.3.0
eventlet>=0.33.0
websockets>=11.0

# GUI and Dashboard
streamlit>=1.25.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Utilities
tqdm>=4.65.0
h5py>=3.9.0
joblib>=1.3.0
PyYAML>=6.0
python-dotenv>=1.0.0
opencv-python>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.0.0
